[svn-r6843] Purpose:     Code cleanup
Description:
    Clean up warnings exposed by compiling on O2K.  Also, revert some of Bill
and my changes to the H5S_mpi_opt_types_g, etc. and settle them back into their
original location.

Platforms tested:
    h5committested.

+++ src/H5RS.c
@@ -392,7 +392,7 @@ H5RS_len(const H5RS_str_t *rs)
     assert(rs);
     assert(rs->s);
 
-    ret_value=HDstrlen(rs->s);
+    ret_value=(ssize_t)HDstrlen(rs->s);
 
 done:
     FUNC_LEAVE_NOAPI(ret_value);
